code/theseus/schema.py
import graphene

import theseus.organizations.schema


class Query(theseus.organizations.schema.Query, graphene.ObjectType):
    # This class will inherit from multiple Queries
    # as we begin to add more apps to our project
    pass

schema = graphene.Schema(query=Query)
