requirements.txt
Django
psycopg2
graphene_django
django-graphiql
django-filter
